+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - The "Updating API" message will now only show on the status bar instead of a popup notification.
+- Instance types will now be named by the Class Name rather than the Instance name
 
 ## [1.2.0] - 2022-06-04